requirements.txt
python-dotenv
pandas
numpy
pymongo[srv]==3.6
seaborn
matplotlib
scikit-learn
certifi
catboost
xgboost
dill
Flask
#-e .